% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_tidy.R
\name{constituency_results_tidy}
\alias{constituency_results_tidy}
\title{constituency_results_tidy}
\usage{
constituency_results_tidy(results, details)
}
\arguments{
\item{results}{The tibble to tidy}

\item{details}{The list to tidy}
}
\description{
constituency_results_tidy
}
