% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_tidy.R
\name{date_tidy}
\alias{date_tidy}
\title{A function that makes date variables returned from the API datable, ie by converting them to POSIXct. Does the same thing for datetimes.}
\usage{
date_tidy(df)
}
\arguments{
\item{df}{The tibble with the undateable dates.}
}
\description{
A function that makes date variables returned from the API datable, ie by converting them to POSIXct. Does the same thing for datetimes.
}
