% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_biog.R
\name{full_biog}
\alias{full_biog}
\title{full_biog}
\usage{
full_biog(ID = NULL, mnis_id = TRUE, ref_dods = FALSE, tidy = TRUE)
}
\arguments{
\item{ID}{The ID number of the member. If left empty, returns an error.}

\item{mnis_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{ref_dods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}

\item{tidy}{Fix the variable names in the data frame to remove '@' characters and superfluous text. Defaults to TRUE.}
}
\description{
Requests all available biographical information for a given member.
}
\examples{
\dontrun{
x <- mnis_full_biog(172)

}
}
\keyword{mnis}
