% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Addresses.R
\name{mnis_Addresses}
\alias{mnis_Addresses}
\title{mnis_Addresses}
\usage{
mnis_Addresses(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Addresses('')

}
}
\keyword{mnis}

