% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mnis_bom}
\alias{mnis_bom}
\title{Strip out BOM from JSON data}
\usage{
mnis_bom(x)
}
\arguments{
\item{x}{The GET return to strip BOM out of}
}
\description{
Strip out BOM from JSON data
}
