% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Committee.R
\name{mnis_Committee}
\alias{mnis_Committee}
\title{mnis_ConstituencyResults}
\usage{
mnis_Committee(committeeID = NULL, former = TRUE)
}
\arguments{
\item{committeeID}{The ID of the committee to return the data for.}

\item{former}{Flag to include both current and former committee members. Defaults to TRUE.}
}
\value{
A list with information on the outcome of the most recent election in a constituency
}
\description{
mnis_ConstituencyResults
}
\examples{
\dontrun{
x <- mnis_ConstituencyResults(3701, 19)

}

}
\keyword{mnis}

