% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Committees.R
\name{mnis_Committees}
\alias{mnis_Committees}
\title{mnis_Committees}
\usage{
mnis_Committees(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Committees(172)

}
}
\keyword{mnis}

