% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Constituencies.R
\name{mnis_Constituencies}
\alias{mnis_Constituencies}
\title{mnis_Constituencies}
\usage{
mnis_Constituencies(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Constituencies(172)

}
}
\keyword{mnis}

