% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_ConstituencyResults.R
\name{mnis_ConstituencyResults}
\alias{mnis_ConstituencyResults}
\title{mnis_ConstituencyResults}
\usage{
mnis_ConstituencyResults(constituencyId = NULL, electionId = 0)
}
\arguments{
\item{constituencyId}{The ID of the constituency to return the data for.}

\item{electionId}{The ID of the election to return the data for. Defaults to 0, which calls the latest result.}
}
\value{
A list with details of the constituency and a data frame with election results.
}
\description{
Returns a list with details of the constituency and a data frame with election results.
}
\examples{
\dontrun{
x <- mnis_ConstituencyResults(constituencyId = 3709, electionId = 0)

}
}
\keyword{mnis}

