% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_defunct.R
\name{mnis_Addresses}
\alias{mnis_Addresses}
\alias{mnis_BasicDetails}
\alias{mnis_BiographyEntries}
\alias{mnis_ConstituencyResults}
\alias{mnis_Committees}
\alias{mnis_Constituencies}
\alias{mnis_ElectionsContested}
\alias{mnis_Experiences}
\alias{mnis_GovernmentPosts}
\alias{mnis_Honours}
\alias{mnis_HouseMemberships}
\alias{mnis_Statuses}
\alias{mnis_Staff}
\alias{mnis_Interests}
\alias{mnis_KnownAs}
\alias{mnis_MaidenSpeeches}
\alias{mnis_OppositionPosts}
\alias{mnis_OtherParliaments}
\alias{mnis_ParliamentaryPosts}
\alias{mnis_Parties}
\alias{mnis_PreferredNames}
\alias{mnis_Department}
\alias{mnis_MemberDate}
\alias{mnis_defunct}
\alias{ref_AddressTypes}
\alias{ref_AnsweringBodies}
\alias{ref_Areas}
\alias{ref_AreaTypes}
\alias{ref_BiographyCategories}
\alias{ref_Cabinets}
\alias{ref_Committees}
\alias{ref_CommitteeTypes}
\alias{ref_Constituencies}
\alias{ref_ConstituencyAreas}
\alias{ref_ConstituencyTypes}
\alias{ref_Countries}
\alias{ref_Departments}
\alias{ref_DisqualificationTypes}
\alias{ref_Elections}
\alias{ref_ElectionTypes}
\alias{ref_EndReasons}
\alias{ref_ExperienceTypes}
\alias{ref_GovernmentPosts}
\alias{ref_GovernmentRanks}
\alias{ref_HonouraryPrefixes}
\alias{ref_HonourLists}
\alias{mnis_GeneralElectionResults}
\alias{ref_GovernmentPostDepartments}
\alias{ref_Honours}
\alias{ref_InterestCategories}
\alias{ref_LordsMembershipTypes}
\alias{ref_LordsRanks}
\alias{ref_OppositionPostDepartments}
\alias{ref_OppositionPosts}
\alias{ref_OppositionRanks}
\alias{ref_OtherParliaments}
\alias{ref_ParliamentaryPosts}
\alias{ref_ParliamentaryRanks}
\alias{ref_ParliamentTypes}
\alias{ref_Parties}
\alias{ref_PartySubTypes}
\alias{ref_PhotoOutputs}
\alias{ref_Statuses}
\alias{ref_Titles}
\title{mnis_defunct}
\description{
Defunct functions for the `mnis` package.
}
