% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_ElectionsContested.R
\name{mnis_ElectionsContested}
\alias{mnis_ElectionsContested}
\title{mnis_ElectionsContested}
\usage{
mnis_ElectionsContested(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_ElectionsContested('')

}
}
\keyword{mnis}

