% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_eligible.R
\name{mnis_eligible}
\alias{mnis_eligible}
\title{mnis_eligible}
\usage{
mnis_eligible(eligible = TRUE, house = "all", party = NULL, tidy = TRUE)
}
\arguments{
\item{eligible}{If the member is currently eligible to sit. Accepts TRUE or FALSE. Defaults to TRUE.}

\item{house}{The house to which the member belongs. Accepts one of 'all', 'lords' and 'commons', defaults to 'all'.}

\item{party}{The party to which a member belongs. Defaults to NULL. The party must be fully spelled out (e.g. 'green party'), the API does not accept searches on this parameter.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to all lower case with underscores between each word. Defaults to TRUE.}
}
\description{
Function returns all members who are eligible to sit in either house. The function requests data in JSON format and parses it to a tibble.
}
\examples{
\dontrun{
x <- mnis_eligible(eligible=FALSE, house='all', party='labour')

x <- mnis_eligible(eligible=TRUE, house='all', party='green party')

}
}
\keyword{mnis}
