% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Experiences.R
\name{mnis_Experiences}
\alias{mnis_Experiences}
\title{mnis_Experiences}
\usage{
mnis_Experiences(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Experiences('')

}
}
\keyword{mnis}

