% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_FullBiog.R
\name{mnis_FullBiog}
\alias{mnis_FullBiog}
\title{mnis_FullBiog}
\usage{
mnis_FullBiog(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Requests all available biographical information for a given member.
}
\examples{
\dontrun{
x <- mnis_FullBiog(172)

}
}
\keyword{mnis}

