% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_GeneralElectionResults.R
\name{mnis_GeneralElectionResults}
\alias{mnis_GeneralElectionResults}
\title{mnis_GeneralElectionResults}
\usage{
mnis_GeneralElectionResults(locationType = NULL, locationName = NULL,
  startDate = NULL, endDate = NULL)
}
\arguments{
\item{locationType}{The ID of the constituency to return the data for.}

\item{locationName}{The ID of the election to return the data for. Call 0 to see the latest result.}

\item{startDate}{Start date of search. In yyyy-mm-dd format.}

\item{endDate}{End date of search. In yyyy-mm-dd format. Defaults to current date if no date is selected.}
}
\value{
A list with information on the outcome of the most recent election in a constituency
}
\description{
mnis_GeneralElectionResults
}
\examples{
\dontrun{
x <- mnis_GeneralElectionResults(locationType = "Country", locationName = "England", startDate = "2010-01-01", endDate = "2016-01-01")

}

}
\keyword{mnis}

