% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Honours.R
\name{mnis_Honours}
\alias{mnis_Honours}
\title{mnis_Honours}
\usage{
mnis_Honours(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Honours(172)

}
}
\keyword{mnis}

