% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_HouseMemberships.R
\name{mnis_HouseMemberships}
\alias{mnis_HouseMemberships}
\title{mnis_HouseMemberships}
\usage{
mnis_HouseMemberships(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_HouseMemberships(172)

}
}
\keyword{mnis}

