% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_KnownAs.R
\name{mnis_KnownAs}
\alias{mnis_KnownAs}
\title{mnis_KnownAs}
\usage{
mnis_KnownAs(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_KnownAs('')

}
}
\keyword{mnis}

