% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_lords_type.R
\name{mnis_lords_type}
\alias{mnis_lords_type}
\title{mnis_lords_type}
\usage{
mnis_lords_type(date = Sys.Date(), tidy = TRUE)
}
\arguments{
\item{date}{A date in yyyy-mm-dd format. Defaults to the current system date.}

\item{tidy}{Fix the variable names in the data frame to remove '@' characters and superfluous text. Defaults to TRUE.}
}
\value{
A data frame with information on the numbers of different types of Lords on a given date.
}
\description{
Calls the API to return a data frame with details on the number of Lords and their affiliations.
}
\examples{
\dontrun{

x <- mnis_lords_type()

}
}
\keyword{mnis}
