% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_MaidenSpeeches.R
\name{mnis_MaidenSpeeches}
\alias{mnis_MaidenSpeeches}
\title{mnis_MaidenSpeeches}
\usage{
mnis_MaidenSpeeches(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_MaidenSpeeches('')

}
}
\keyword{mnis}

