% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_MemberDate.R
\name{mnis_MemberDate}
\alias{mnis_MemberDate}
\title{mnis_MemberDate}
\usage{
mnis_MemberDate(ID = NULL, Date = NULL)
}
\arguments{
\item{ID}{The ID of the member.}

\item{Date}{A date in yyyy-mm-dd format. Defaults to the current system date.}
}
\value{
Returns a data frame with the given member's status on the given date.
}
\description{
Returns a data frame with a members status on a given date.
}
\examples{
\dontrun{
x <- mnis_MemberDate(172)

}
}
\keyword{mnis}

