% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_OppositionPosts.R
\name{mnis_OppositionPosts}
\alias{mnis_OppositionPosts}
\title{mnis_OppositionPosts}
\usage{
mnis_OppositionPosts(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_OppositionPosts('')

}
}
\keyword{mnis}

