% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_OtherParliaments.R
\name{mnis_OtherParliaments}
\alias{mnis_OtherParliaments}
\title{mnis_OtherParliaments}
\usage{
mnis_OtherParliaments(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_OtherParliaments(172)

}
}
\keyword{mnis}

