% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_ParliamentaryPosts.R
\name{mnis_ParliamentaryPosts}
\alias{mnis_ParliamentaryPosts}
\title{mnis_ParliamentaryPosts}
\usage{
mnis_ParliamentaryPosts(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_ParliamentaryPosts(172)

}
}
\keyword{mnis}

