% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Parties.R
\name{mnis_Parties}
\alias{mnis_Parties}
\title{mnis_Parties}
\usage{
mnis_Parties(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Parties('')

}
}
\keyword{mnis}

