% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_PartyState.R
\name{mnis_PartyState}
\alias{mnis_PartyState}
\title{mnis_PartyState}
\usage{
mnis_PartyState(House = "Commons", Date = Sys.Date())
}
\arguments{
\item{House}{The house of parliament. Defaults to 'Commons'.}

\item{Date}{A date in yyyy-mm-dd format. Defaults to the current date.}
}
\value{
A data frame with information on the numbers and gender of MPs, by party, by party, for the given date.
}
\description{
A data frame with information on the numbers and gender of MPs, by party, for the given date.
}
\examples{
\dontrun{
x <- mnis_PartyState('2012-01-12')

}
}
\keyword{mnis}

