% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_PreferredNames.R
\name{mnis_PreferredNames}
\alias{mnis_PreferredNames}
\title{mnis_PreferredNames}
\usage{
mnis_PreferredNames(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Lookup function for the MNIS API. The function requests data in JSON format and parses it to a data frame containing the prefered name of the selected member.
}
\examples{
\dontrun{
x <- mnis_PreferredNames(172)

}
}
\keyword{mnis}

