% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Staff.R
\name{mnis_Staff}
\alias{mnis_Staff}
\title{mnis_Staff}
\usage{
mnis_Staff(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Staff('')

}
}
\keyword{mnis}

