% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Statuses.R
\name{mnis_Statuses}
\alias{mnis_Statuses}
\title{mnis_Statuses}
\usage{
mnis_Statuses(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Statuses('')

}
}
\keyword{mnis}

