% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_Additional.R
\name{mp_BasicDetails}
\alias{mp_Addresses}
\alias{mp_BasicDetails}
\alias{mp_BiographyEntries}
\alias{mp_Committees}
\alias{mp_Constituencies}
\alias{mp_ElectionsContested}
\alias{mp_Experiences}
\alias{mp_GovernmentPosts}
\alias{mp_Honours}
\alias{mp_HouseMemberships}
\alias{mp_Interests}
\alias{mp_KnownAs}
\alias{mp_MaidenSpeeches}
\alias{mp_OppositionPosts}
\alias{mp_OtherParliaments}
\alias{mp_ParliamentaryPosts}
\alias{mp_Parties}
\alias{mp_PreferredNames}
\alias{mp_Staff}
\alias{mp_Statuses}
\title{mp_additional}
\usage{
mp_BasicDetails(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_BiographyEntries(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Committees(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Addresses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Constituencies(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_ElectionsContested(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Experiences(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_GovernmentPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Honours(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_HouseMemberships(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Statuses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Staff(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Interests(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_KnownAs(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_MaidenSpeeches(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_OppositionPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_OtherParliaments(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_ParliamentaryPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_Parties(ID = NULL, mem_id = TRUE, refDods = FALSE)

mp_PreferredNames(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame. Variable descriptions are taken from the mnis website.
}
\examples{
\dontrun{
x <- mp_BasicDetails(172)

}
}
\seealso{
\code{\link{mp_FullBiog}} \code{\link{mp_Extra}}
}
\keyword{mnis}

