% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_FullBiog.R
\name{mp_FullBiog}
\alias{mp_FullBiog}
\title{mp_FullBiog}
\usage{
mp_FullBiog(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mp_FullBiog(172)

}
}
\keyword{mnis}

