% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{sans_bom}
\alias{sans_bom}
\title{Tests whether a raw httr response or character vector has a byte order mark (BOM)}
\usage{
sans_bom(resp)
}
\arguments{
\item{resp}{\code{httr} response object or a character vector}
}
\value{
Converted content
}
\description{
Tests whether a raw httr response or character vector has a byte order mark (BOM)
}
\references{
\href{http://www.unicode.org/faq/utf_bom.html}{UTF-8, UTF-16, UTF-32 & BOM}
}
\author{
@hrbrmstr
}
