% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_tidy.R
\name{tidy_bom}
\alias{tidy_bom}
\title{A function to strip Byte Order Marks (BOM) out of JSON data returned from the API.}
\usage{
tidy_bom(df)
}
\arguments{
\item{df}{The GET returned from call to API.}
}
\description{
A function to strip Byte Order Marks (BOM) out of JSON data returned from the API.
}
