% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_additional.R
\name{mnis_additional}
\alias{mnis_Addresses}
\alias{mnis_BasicDetails}
\alias{mnis_BiographyEntries}
\alias{mnis_Committees}
\alias{mnis_Constituencies}
\alias{mnis_ElectionsContested}
\alias{mnis_Experiences}
\alias{mnis_GovernmentPosts}
\alias{mnis_Honours}
\alias{mnis_HouseMemberships}
\alias{mnis_Interests}
\alias{mnis_KnownAs}
\alias{mnis_MaidenSpeeches}
\alias{mnis_OppositionPosts}
\alias{mnis_OtherParliaments}
\alias{mnis_ParliamentaryPosts}
\alias{mnis_Parties}
\alias{mnis_PreferredNames}
\alias{mnis_Staff}
\alias{mnis_Statuses}
\alias{mnis_additional}
\alias{mnis_addresses}
\alias{mnis_basic_details}
\alias{mnis_biography_entries}
\alias{mnis_committees}
\alias{mnis_constituencies}
\alias{mnis_elections_contested}
\alias{mnis_experiences}
\alias{mnis_government_posts}
\alias{mnis_honours}
\alias{mnis_house_memberships}
\alias{mnis_interests}
\alias{mnis_known_as}
\alias{mnis_maiden_speeches}
\alias{mnis_opposition_posts}
\alias{mnis_other_parliaments}
\alias{mnis_parliamentary_posts}
\alias{mnis_parties}
\alias{mnis_preferred_names}
\alias{mnis_staff}
\alias{mnis_statuses}
\title{mnis_additional}
\usage{
mnis_additional()

mnis_basic_details(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_biography_entries(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_committees(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_addresses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_constituencies(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_elections_contested(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_experiences(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_government_posts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_honours(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_house_memberships(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_statuses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_staff(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_interests(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_known_as(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_maiden_speeches(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_opposition_posts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_other_parliaments(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_parliamentary_posts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_parties(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_preferred_names(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame. Variable descriptions are taken from the mnis website.
}
\examples{
\dontrun{
x <- mnis_basic_details(172)

}
}
\seealso{
\code{\link{mnis_full_biog}} \code{\link{mnis_extra}}
}
\keyword{mnis}

