% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Additional.R
\name{mnis_Additional}
\alias{mnis_Additional}
\alias{mnis_Addresses}
\alias{mnis_BasicDetails}
\alias{mnis_BiographyEntries}
\alias{mnis_Committees}
\alias{mnis_Constituencies}
\alias{mnis_ElectionsContested}
\alias{mnis_Experiences}
\alias{mnis_GovernmentPosts}
\alias{mnis_Honours}
\alias{mnis_HouseMemberships}
\alias{mnis_Interests}
\alias{mnis_KnownAs}
\alias{mnis_MaidenSpeeches}
\alias{mnis_OppositionPosts}
\alias{mnis_OtherParliaments}
\alias{mnis_ParliamentaryPosts}
\alias{mnis_Parties}
\alias{mnis_PreferredNames}
\alias{mnis_Staff}
\alias{mnis_Statuses}
\title{mnis_additional}
\usage{
mnis_Additional()

mnis_BasicDetails(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_BiographyEntries(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Committees(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Addresses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Constituencies(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_ElectionsContested(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Experiences(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_GovernmentPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Honours(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_HouseMemberships(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Statuses(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Staff(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Interests(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_KnownAs(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_MaidenSpeeches(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_OppositionPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_OtherParliaments(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_ParliamentaryPosts(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_Parties(ID = NULL, mem_id = TRUE, refDods = FALSE)

mnis_PreferredNames(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame. Variable descriptions are taken from the mnis website.
}
\examples{
\dontrun{
x <- mnis_BasicDetails(172)

}
}
\seealso{
\code{\link{mnis_FullBiog}} \code{\link{mnis_Extra}}
}
\keyword{mnis}

