% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_additional.R
\name{mnis_additional}
\alias{mnis_additional}
\alias{mnis_basic_details}
\alias{mnis_biography_entries}
\alias{mnis_committees}
\alias{mnis_addresses}
\alias{mnis_constituencies}
\alias{mnis_elections_contested}
\alias{mnis_experiences}
\alias{mnis_government_posts}
\alias{mnis_honours}
\alias{mnis_house_memberships}
\alias{mnis_statuses}
\alias{mnis_staff}
\alias{mnis_interests}
\alias{mnis_known_as}
\alias{mnis_maiden_speeches}
\alias{mnis_opposition_posts}
\alias{mnis_other_parliaments}
\alias{mnis_parliamentary_posts}
\alias{mnis_parties}
\alias{mnis_preferred_names}
\title{mnis_additional}
\usage{
mnis_additional()

mnis_basic_details(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_biography_entries(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_committees(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_addresses(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_constituencies(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_elections_contested(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_experiences(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_government_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_honours(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_house_memberships(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_statuses(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_staff(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_interests(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_known_as(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_maiden_speeches(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_opposition_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_other_parliaments(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_parliamentary_posts(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_parties(ID = NULL, ref_dods = FALSE, tidy = TRUE)

mnis_preferred_names(ID = NULL, ref_dods = FALSE, tidy = TRUE)
}
\arguments{
\item{ID}{The member ID value. If empty, function calls \code{\link{mnis_all_members}} and returns basic information on all members of both houses.}

\item{ref_dods}{Request based on the DODS membership ID scheme. Defaults to FALSE, where it requests data based on the default membership ID scheme.}

\item{tidy}{Fix the variable names in the tibble to remove '@' characters and superfluous text. Defaults to TRUE.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a tibble.
}
\examples{
\dontrun{

x <- mnis_basic_details(172)

}
\dontrun{

x <- mnis_additional()

}
}
\seealso{
\code{\link{mnis_full_biog}} \code{\link{mnis_extra}}
}
\keyword{mnis}
