% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Addresses.R
\name{mnis_Addresses}
\alias{mnis_Addresses}
\title{mnis_Addresses}
\usage{
mnis_Addresses(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}
}
\description{
Requests addresses for a given member. The API returns address of parliamentary offices, constituency offices, websites and social media accounts. The function requests data in JSON format and parses it to a data frame. Only one of mem_id, refDods and pims_id can be TRUE.
}
\examples{
\dontrun{
x <- mnis_Addresses(172)

}
}
\keyword{mnis}

