% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_all_members.R
\name{mnis_all_members}
\alias{mnis_all_members}
\title{mnis_all_members}
\usage{
mnis_all_members(house = "all", party = NULL, joined_since = NULL,
  tidy = TRUE)
}
\arguments{
\item{house}{The house to which the member belongs. Accepts one of 'all', 'lords' and 'commons', defaults to 'all'.}

\item{party}{The party to which a member belongs. Defaults to NULL.}

\item{joined_since}{All members who joined after a given date, in yyyy-mm-dd format.}

\item{tidy}{Fix the variable names in the tibble to remove special characters and superfluous text, and converts the variable names to all lower case with underscores between each word. Defaults to TRUE.}
}
\value{
A tibble with all members of both houses
}
\description{
mnis_all_members
}
\examples{
\dontrun{
x <- mnis_all_members(house = 'all', party = NULL, joined_since = NULL, tidy = TRUE)
}

}
\keyword{mnis}
