% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_all_reference.R
\name{mnis_all_reference}
\alias{mnis_all_reference}
\title{Returns a list of 39 tibbles, with the name and data returned from each respective reference function.}
\usage{
mnis_all_reference(tidy = TRUE, tidy_style = "snake_case")
}
\arguments{
\item{tidy}{Fix the variable names in each tibble to remove special characters and superfluous text, and convert all the variable names to snake_case. Defaults to TRUE.}

\item{tidy_style}{The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".}
}
\value{
Returns a list of 39 tibbles, with the name and data returned from each respective reference function.
}
\description{
Returns a list of 39 tibbles, with the name and data returned from each respective reference function.
}
\examples{
\dontrun{

z <- mnis_all_reference()

}
}
\seealso{
\code{\link{mnis_reference}}
}
