% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_all_reference.R
\name{mnis_all_reference}
\alias{mnis_all_reference}
\title{mnis_all_reference}
\usage{
mnis_all_reference(tidy = TRUE)
}
\arguments{
\item{tidy}{Fix the variable names in each tibble to remove special characters and superfluous text, and convert all the variable names to snake_case. Defaults to TRUE.}
}
\value{
Returns a list of 39 tibbles, with the name and data returned from each respective reference function.
}
\description{
mnis_all_reference
}
\examples{
\dontrun{

mnis_all_reference()

}
}
