% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_base.R
\name{mnis_base}
\alias{mnis_base}
\title{A generic function for the MNIS API lookup. The function requests data in JSON format, but the type of object, and all URLs, paths and parameters are user-defined. \code{mnis_base} does not include the option to tidy variable names and data types, but the \code{\link{mnis_tidy}} function can be called on the data returned from the API.}
\usage{
mnis_base(request)
}
\arguments{
\item{request}{The request query being made to the MNIS URL}
}
\description{
A generic function for the MNIS API lookup. The function requests data in JSON format, but the type of object, and all URLs, paths and parameters are user-defined. \code{mnis_base} does not include the option to tidy variable names and data types, but the \code{\link{mnis_tidy}} function can be called on the data returned from the API.
}
\examples{
\dontrun{

x <- mnis_base('House=Commons|IsEligible=true/')

}
}
\keyword{mnis}
