% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_base.R
\name{mnis_base}
\alias{mnis_base}
\title{mnis_Base}
\usage{
mnis_base(request)
}
\arguments{
\item{request}{The request query being made to the MNIS URL}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Base('House=Commons|IsEligible=true/')

}
}
\keyword{mnis}

