% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_base.R
\name{mnis_base}
\alias{mnis_base}
\title{mnis_base}
\usage{
mnis_base(request)
}
\arguments{
\item{request}{The request query being made to the MNIS URL}
}
\description{
A generic function for the MNIS API lookup. The function requests data in JSON format, but the type of object, and all URLs, paths and parameters are user-defined.
}
\examples{
\dontrun{

x <- mnis_base('House=Commons|IsEligible=true/')

}
}
\keyword{mnis}
