% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_BasicDetails.R
\name{mnis_BasicDetails}
\alias{mnis_BasicDetails}
\title{mnis_BasicDetails}
\usage{
mnis_BasicDetails(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_BasicDetails(172)

}
}
\keyword{mnis}

