% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_BiographyEntries.R
\name{mnis_BiographyEntries}
\alias{mnis_BiographyEntries}
\title{mnis_Addresses}
\usage{
mnis_BiographyEntries(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_BiographyEntries('')

}
}
\keyword{mnis}

