% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Committees.R
\name{mnis_Committees}
\alias{mnis_Committees}
\title{mnis_Committees}
\usage{
mnis_Committees(ID = NULL, mem_id = TRUE, refDods = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the Dods membership ID scheme. Defaults to FALSE.}
}
\description{
Requests data on committee membership. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Committees(172)

}
}
\keyword{mnis}

