% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_constituency_results.R
\name{mnis_constituency_results}
\alias{mnis_constituency_results}
\title{Returns a single object containing a list with details of the constituency and a tibble with election results. For constituency IDs, see \code{\link{ref_constituencies}}.}
\usage{
mnis_constituency_results(constituency_id = NULL, election_id = 0,
  tidy = TRUE, tidy_style = "snake_case")
}
\arguments{
\item{constituency_id}{The ID of the constituency to return the data for. This parameter cannot be empty.}

\item{election_id}{The ID of the election to return the data for. Defaults to 0, which calls the most recent result, either the result of the last general election, or the result of the last byelection held since that election.}

\item{tidy}{If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.}

\item{tidy_style}{The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".}
}
\value{
A list with details of the constituency, labelled 'details' and a tibble with election results, labelled 'results'. The list and tibble are stored in a single object.
}
\description{
Returns a single object containing a list with details of the constituency and a tibble with election results. For constituency IDs, see \code{\link{ref_constituencies}}.
}
\examples{
\dontrun{

x <- mnis_constituency_results(constituency_id = 3709, election_id = 0)

}
}
\seealso{
\code{\link{mnis_reference}}
}
\keyword{mnis}
