% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_constituency_results.R
\name{mnis_constituency_results}
\alias{mnis_constituency_results}
\alias{mnis_ConstituencyResults}
\title{mnis_constituency_results}
\usage{
mnis_constituency_results(constituencyId = NULL, electionId = 0,
  tidy = TRUE)
}
\arguments{
\item{constituencyId}{The ID of the constituency to return the data for.}

\item{electionId}{The ID of the election to return the data for. Defaults to 0, which calls the latest result.}

\item{tidy}{Fix the variable names in the data frame to remove '@' characters and superfluous text. Defaults to TRUE.}
}
\value{
A list with details of the constituency, labelled 'details' and a data frame with election results, labelled 'results'. The list and data frame are stored in a single object.
}
\description{
Returns a list with details of the constituency and a data frame with election results.
}
\examples{
\dontrun{
x <- mnis_constituency_results(constituencyId = 3709, electionId = 0)

}
}
\keyword{mnis}
