% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_ConstituencyResults.R
\name{mnis_ConstituencyResults}
\alias{mnis_ConstituencyResults}
\title{mnis_ConstituencyResults}
\usage{
mnis_ConstituencyResults(constituencyId = NULL, electionId = NULL)
}
\arguments{
\item{constituencyId}{The ID of the constituency to return the data for.}

\item{electionId}{The ID of the election to return the data for. Call 0 to see the latest result.}
}
\value{
A list with information on the outcome of the most recent election in a constituency
}
\description{
mnis_ConstituencyResults
}
\examples{
\dontrun{
x <- mnis_ConstituencyResults(3701, 19)

}

}
\keyword{mnis}

