% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_defunct.R
\name{mnis_Addresses}
\alias{mnis_Addresses}
\alias{mnis_BasicDetails}
\alias{mnis_BiographyEntries}
\alias{mnis_Committees}
\alias{mnis_Constituencies}
\alias{mnis_ElectionsContested}
\alias{mnis_Experiences}
\alias{mnis_GovernmentPosts}
\alias{mnis_Honours}
\alias{mnis_HouseMemberships}
\alias{mnis_Statuses}
\alias{mnis_Staff}
\alias{mnis_Interests}
\alias{mnis_KnownAs}
\alias{mnis_MaidenSpeeches}
\alias{mnis_OppositionPosts}
\alias{mnis_OtherParliaments}
\alias{mnis_ParliamentaryPosts}
\alias{mnis_Parties}
\alias{mnis_PreferredNames}
\alias{mnis_Department}
\alias{mnis_MemberDate}
\title{mnis_defunct}
\description{
Defunct functions for the `mnis` package.
}
