% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_department.R
\name{mnis_department}
\alias{mnis_department}
\title{Request data on the holders of cabinet/shadow cabinet positions. Request specific departments by department ID (see \code{\link{mnis_reference}} for the \code{ref_department} function to retrieve departmental IDs).}
\usage{
mnis_department(department_id = 0, bench = "Government", former = TRUE,
  tidy = TRUE, tidy_style = "snake_case")
}
\arguments{
\item{department_id}{The department look up. 0 returns the cabinet/shadow cabinet, -1 returns a list of all ministers. Defaults to 0.}

\item{bench}{Flag to return either Government or Opposition information. Defaults to 'Government'. The API is case sensitive on this parameter, so 'Government' or 'Opposition' will work, but 'government' and 'opposition' will not.}

\item{former}{Flag to include both current and former ministers/shadow ministers. Defaults to TRUE. If FALSE, only includes current ministers/shadow ministers.}

\item{tidy}{If TRUE, fixes the variable names in the tibble to remove non-alphanumeric characters and superfluous text, and convert to a consistent style. Defaults to TRUE.}

\item{tidy_style}{The style to convert variable names to, if tidy=TRUE. Accepts one of "snake_case", "camelCase" and "period.case". Defaults to "snake_case".}
}
\value{
A tibble with information on departments and ministers/shadow ministers.
}
\description{
Request data on the holders of cabinet/shadow cabinet positions. Request specific departments by department ID (see \code{\link{mnis_reference}} for the \code{ref_department} function to retrieve departmental IDs).
}
\examples{
\dontrun{

x <- mnis_department(department_id = 0, bench = 'Government', former=TRUE)

}

}
\seealso{
\code{\link{mnis_reference}}
}
\keyword{mnis}
