% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_department.R
\name{mnis_department}
\alias{mnis_department}
\title{mnis_department}
\usage{
mnis_department(departmentId = 0, bench = "Government", former = TRUE)
}
\arguments{
\item{departmentId}{The department look up. 0 returns the cabinet/shadow cabinet, -1 returns a list of all ministers/}

\item{bench}{Flag to return either Government or Opposition information. Defaults to Government.}

\item{former}{Flag to include both current and former ministers/shadow ministers. Defaults to TRUE.}
}
\value{
A list with information on the outcome of the most recent election in a constituency
}
\description{
mnis_department
}
\examples{
\dontrun{
x <- mnis_department(departmentId = 0, bench = 'Government', former=TRUE)

}

}
\keyword{mnis}

