% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_extra.R, R/mnis_extra_old.R
\name{mnis_extra}
\alias{mnis_extra}
\title{mnis_extra}
\usage{
mnis_extra(ID, FullBiog = FALSE, Addresses = FALSE, BasicDetails = FALSE,
  BiographyEntries = FALSE, Committees = FALSE, Constituencies = FALSE,
  ElectionsContested = FALSE, Experiences = FALSE,
  GovernmentPosts = FALSE, Honours = FALSE, HouseMemberships = FALSE,
  Interests = FALSE, KnownAs = FALSE, MaidenSpeeches = FALSE,
  OppositionPosts = FALSE, OtherParliaments = FALSE,
  ParliamentaryPosts = FALSE, Parties = FALSE, PreferredNames = FALSE,
  Staff = FALSE, Statuses = FALSE)

mnis_extra(ID, FullBiog = FALSE, Addresses = FALSE, BasicDetails = FALSE,
  BiographyEntries = FALSE, Committees = FALSE, Constituencies = FALSE,
  ElectionsContested = FALSE, Experiences = FALSE,
  GovernmentPosts = FALSE, Honours = FALSE, HouseMemberships = FALSE,
  Interests = FALSE, KnownAs = FALSE, MaidenSpeeches = FALSE,
  OppositionPosts = FALSE, OtherParliaments = FALSE,
  ParliamentaryPosts = FALSE, Parties = FALSE, PreferredNames = FALSE,
  Staff = FALSE, Statuses = FALSE)
}
\arguments{
\item{ID}{The ID number of the member. Defaults to NULL.}

\item{FullBiog}{Returns all data held about a Member. This will only be permitted on queries that return five records or fewer. All other output types return subsets of this data. Defaults to TRUE.}

\item{Addresses}{Member address information (e.g. website, twitter, consituency address etc...). Defaults to TRUE.}

\item{BasicDetails}{Basic information about the Member (e.g. given name, HoL membership type, oaths etc...) Defaults to TRUE.}

\item{BiographyEntries}{Member biographical information (e.g. countries of interest, policy expertise etc...) Defaults to TRUE.}

\item{Committees}{Committees a Member sits or has sat on as well details on committee chairing. Defaults to TRUE.}

\item{Constituencies}{Constituencies a Member has represented. Defaults to TRUE.}

\item{ElectionsContested}{Elections a Member has contested but not won. Defaults to TRUE.}

\item{Experiences}{Non-parliamentary experience of a Member. Defaults to TRUE.}

\item{GovernmentPosts}{Government posts a Member has held. Defaults to TRUE.}

\item{Honours}{Honours (e.g. MBE, OBE etc...) held by a Member. Defaults to TRUE.}

\item{HouseMemberships}{House membership list of a Member. Defaults to TRUE.}

\item{Interests}{Registered interests (financial) of a Member. Defaults to TRUE.}

\item{KnownAs}{Details of names a Member has chosen to be known as instead of their full title (HoL members only). Defaults to TRUE.}

\item{MaidenSpeeches}{Maiden speech dates for a Member. Defaults to TRUE.}

\item{OppositionPosts}{Opposition posts a Member has held. Defaults to TRUE.}

\item{OtherParliaments}{Other Parliaments that a Member has held a membership of. Defaults to TRUE.}

\item{ParliamentaryPosts}{Parliamentary posts a Member has held. Defaults to TRUE.}

\item{Parties}{Party affiliations of a Member. Defaults to TRUE.}

\item{PreferredNames}{Full set of data about a Members' name (e.g. surname, forename, Honorary prefixes, full details of HoL title and rank etc...). Defaults to TRUE.}

\item{Staff}{Staff employed by a Member. Defaults to TRUE.}

\item{Statuses}{Status history (e.g. suspensions and disqualifications) for a Member. Defaults to TRUE.}

\item{mem_id}{Request based on the default membership ID scheme. Defaults to TRUE.}

\item{refDods}{Request based on the DODS membership ID scheme. Defaults to FALSE.}

\item{pims_id}{Request based on the PIMS membership ID scheme. Defaults to FALSE.}

\item{ID}{The ID number}

\item{FullBiog}{Returns all data held about a Member. This will only be permitted on queries that return five records or fewer. All other output types return subsets of this data.}

\item{Addresses}{Member address information (e.g. website, twitter, consituency address etc...)}

\item{BasicDetails}{Basic information about the Member (e.g. given name, HoL membership type, oaths etc...).}

\item{BiographyEntries}{Member biographical information (e.g. countries of interest, policy expertise etc...).}

\item{Committees}{Committees a Member sits or has sat on as well details on committee chairing.}

\item{Constituencies}{Constituencies a Member has represented.}

\item{ElectionsContested}{Elections a Member has contested but not won.}

\item{Experiences}{Non-parliamentary experience of a Member.}

\item{GovernmentPosts}{Government posts a Member has held.}

\item{Honours}{Honours (e.g. MBE, OBE etc...) held by a Member.}

\item{HouseMemberships}{House membership list of a Member.}

\item{Interests}{Registered interests (financial) of a Member.}

\item{KnownAs}{Details of names a Member has chosen to be known as instead of their full title (HoL members only).}

\item{MaidenSpeeches}{Maiden speech dates for a Member.}

\item{OppositionPosts}{Opposition posts a Member has held.}

\item{OtherParliaments}{Other Parliaments that a Member has held a membership of.}

\item{ParliamentaryPosts}{Parliamentary posts a Member has held.}

\item{Parties}{Party affiliations of a Member.}

\item{PreferredNames}{Full set of data about a Members' name (e.g. surname, forename, Honorary prefixes, full details of HoL title and rank etc...)}

\item{Staff}{Staff employed by a Member.}

\item{Statuses}{Status history (e.g. suspensions and disqualifications) for a Member.}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame. Variable descriptions are taken from the mnis website.

Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_extra(172)

}
\dontrun{
x <- mnis_extra('House=Commons|IsEligible=true/')

}
}
\keyword{mnis}

