% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_FullBiog.R
\name{mnis_FullBiog}
\alias{mnis_FullBiog}
\title{mnis_FullBiog}
\usage{
mnis_FullBiog(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_FullBiog(172)

}
}
\keyword{mnis}

