% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Honours.R
\name{mnis_Honours}
\alias{mnis_Honours}
\title{mnis_Honours}
\usage{
mnis_Honours(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame.
}
\examples{
\dontrun{
x <- mnis_Honours('')

}
}
\keyword{mnis}

