% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_LordsType.R
\name{mnis_LordsType}
\alias{mnis_LordsType}
\title{mnis_LordsType}
\usage{
mnis_LordsType(Date = NULL)
}
\arguments{
\item{Date}{A date in yyyy-mm-dd format}
}
\value{
A data frame with information on the numbers of different types of Lords
}
\description{
mnis_LordsType
}
\examples{
\dontrun{
x <- mnis_LordsType()

}

}
\keyword{mnis}

