% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_member_date.R
\name{mnis_member_date}
\alias{mnis_member_date}
\alias{mnis_MemberDate}
\title{mnis_member_date}
\usage{
mnis_member_date(ID = NULL, Date = Sys.Date(), tidy = TRUE)
}
\arguments{
\item{ID}{The ID of the member.}

\item{Date}{A date in yyyy-mm-dd format. Defaults to the current system date.}

\item{tidy}{Fix the variable names in the data frame to remove special characters and superfluous text, and converts the variable names to all lower case with underscores between each word. Defaults to TRUE.}
}
\value{
Returns a data frame with the given member's status on the given date.
}
\description{
Returns a data frame with a members status on a given date.
}
\examples{
\dontrun{
x <- mnis_member_date(172)

}
}
\keyword{mnis}
