% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_MemberDate.R
\name{mnis_MemberDate}
\alias{mnis_MemberDate}
\title{mnis_MemberDate}
\usage{
mnis_MemberDate(ID = NULL, Date = NULL)
}
\arguments{
\item{ID}{The house of parliament. Defaults to "Commons".}

\item{Date}{A date in yyyy-mm-dd format}
}
\value{
This will return a data frame for the specified member id on the date given - useful for knowing how a member should have been addressed in historical applications.
}
\description{
mnis_MemberDate
}
\examples{
\dontrun{
x <- mnis_MemberDate()

}

}
\keyword{mnis}

