% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_mps_on_date.R
\name{mnis_mps_on_date}
\alias{mnis_mps_on_date}
\title{mnis_mps_on_date}
\usage{
mnis_mps_on_date(date = Sys.Date(), tidy = TRUE)
}
\arguments{
\item{date}{The date to return the list of mps from. Defaults to current system date.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\value{
A tibble with all members of the House of Commons eligible to sit on a given date.
}
\description{
mnis_mps_on_date
}
\examples{
\dontrun{

x <- mnis_mps_on_date()

}
}
