% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_party_state.R
\name{mnis_party_state}
\alias{mnis_party_state}
\title{mnis_party_state}
\usage{
mnis_party_state(House = "Commons", Date = Sys.Date(), tidy = TRUE)
}
\arguments{
\item{House}{The house of parliament. Defaults to 'Commons'.}

\item{Date}{A date in yyyy-mm-dd format. Defaults to the current system date.}

\item{tidy}{Fix the variable names in the data frame to remove special characters and superfluous text, and converts the variable names to all lower case with underscores between each word. Defaults to TRUE.}
}
\value{
A data frame with information on the numbers and gender of MPs, by party, by party, for the given date.
}
\description{
A data frame with information on the numbers and gender of MPs, by party, for the given date.
}
\examples{
\dontrun{
x <- mnis_party_state('2012-01-12')

}
}
\keyword{mnis}
