% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_peers_on_date.R
\name{mnis_peers_on_date}
\alias{mnis_peers_on_date}
\title{mnis_peers_on_date}
\usage{
mnis_peers_on_date(date1 = Sys.Date(), date2 = NULL, tidy = TRUE)
}
\arguments{
\item{date1}{The date to return the list of peers from. Defaults to current system date. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}.}

\item{date2}{An optional query parameter. Accepts character values in "YYYY-MM-DD" format, and objects of class Date, POSIXt, POSIXct, POSIXlt or anything else than can be coerced to a date with \code{as.Date()}. If not NULL, the function returns a list of all peers in the House of Lords between date2 and date1. Defaults to NULL.}

\item{tidy}{Fix the variable names in the tibble to remove extra characters, superfluous text and convert variable names to snake_case. Defaults to TRUE.}
}
\value{
A tibble with information on all Peers who were members of the House of Lords on the date specificed (if only date1 is included as a parameter), or on or between the two dates if both date1 and date2 are specified.
}
\description{
mnis_peers_on_date
}
\examples{
\dontrun{

x <- mnis_peers_on_date()

}
}
