% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_PreferredNames.R
\name{mnis_PreferredNames}
\alias{mnis_PreferredNames}
\title{mnis_PreferredNames}
\usage{
mnis_PreferredNames(ID = NULL)
}
\arguments{
\item{ID}{The ID number}
}
\description{
Lookup function for the MNIS API. The function requests data in JSON format and parses it to a data frame containing the prefered name of the selected member.
}
\examples{
\dontrun{
x <- mnis_PreferredNames(172)

}
}
\keyword{mnis}

