% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Reference.R
\name{ref_AddressTypes}
\alias{ref_AddressTypes}
\alias{ref_AnsweringBodies}
\alias{ref_AreaTypes}
\alias{ref_Areas}
\alias{ref_BiographyCategories}
\alias{ref_Cabinets}
\alias{ref_CommitteeTypes}
\alias{ref_Committees}
\alias{ref_Constituencies}
\alias{ref_ConstituencyAreas}
\alias{ref_ConstituencyTypes}
\alias{ref_Countries}
\alias{ref_Departments}
\alias{ref_DisqualificationTypes}
\alias{ref_ElectionTypes}
\alias{ref_Elections}
\alias{ref_EndReasons}
\alias{ref_ExperienceTypes}
\alias{ref_GovernmentPostDepartments}
\alias{ref_GovernmentPosts}
\alias{ref_GovernmentRanks}
\alias{ref_HonourLists}
\alias{ref_HonouraryPrefixes}
\alias{ref_Honours}
\alias{ref_InterestCategories}
\alias{ref_LordsMembershipTypes}
\alias{ref_LordsRanks}
\alias{ref_OppositionPostDepartments}
\alias{ref_OppositionPosts}
\alias{ref_OppositionRanks}
\alias{ref_OtherParliaments}
\alias{ref_ParliamentTypes}
\alias{ref_ParliamentaryPosts}
\alias{ref_ParliamentaryRanks}
\alias{ref_Parties}
\alias{ref_PartySubTypes}
\alias{ref_PhotoOutputs}
\alias{ref_Statuses}
\alias{ref_titles}
\title{mnis_Reference}
\usage{
ref_AddressTypes(x)

ref_AnsweringBodies(x)

ref_Areas(x)

ref_AreaTypes(x)

ref_BiographyCategories(x)

ref_Cabinets(x)

ref_Committees(x)

ref_CommitteeTypes(x)

ref_Constituencies(x)

ref_ConstituencyAreas(x)

ref_ConstituencyTypes(x)

ref_Countries(x)

ref_Departments(x)

ref_DisqualificationTypes(x)

ref_Elections(x)

ref_ElectionTypes(x)

ref_EndReasons(x)

ref_ExperienceTypes(x)

ref_GovernmentPostDepartments(x)

ref_GovernmentPosts(x)

ref_GovernmentRanks(x)

ref_HonouraryPrefixes(x)

ref_HonourLists(x)

ref_Honours(x)

ref_InterestCategories(x)

ref_LordsMembershipTypes(x)

ref_LordsRanks(x)

ref_OppositionPostDepartments(x)

ref_OppositionPosts(x)

ref_OppositionRanks(x)

ref_OtherParliaments(x)

ref_ParliamentaryPosts(x)

ref_ParliamentaryRanks(x)

ref_ParliamentTypes(x)

ref_Parties(x)

ref_PartySubTypes(x)

ref_PhotoOutputs(x)

ref_Statuses(x)

ref_titles(x)
}
\arguments{
\item{x}{Basic argument to function call}
}
\description{
Basic function for the MNIS API lookup. The function requests data in JSON format and parses it to a data frame. Variable descriptions are taken from the mnis website.
}
\examples{
\dontrun{
x <- ref_AddressTypes()

x <- ref_AnsweringBodies()

}

}
\keyword{mnis}

