% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_Reference.R
\name{ref_AddressTypes}
\alias{ref_AddressTypes}
\alias{ref_AnsweringBodies}
\alias{ref_AreaTypes}
\alias{ref_Areas}
\alias{ref_BiographyCategories}
\alias{ref_Cabinets}
\alias{ref_CommitteeTypes}
\alias{ref_Committees}
\alias{ref_Constituencies}
\alias{ref_ConstituencyAreas}
\alias{ref_ConstituencyTypes}
\alias{ref_Countries}
\alias{ref_Departments}
\alias{ref_DisqualificationTypes}
\alias{ref_ElectionTypes}
\alias{ref_Elections}
\alias{ref_EndReasons}
\alias{ref_ExperienceTypes}
\alias{ref_GovernmentPostDepartments}
\alias{ref_GovernmentPosts}
\alias{ref_GovernmentRanks}
\alias{ref_HonourLists}
\alias{ref_HonouraryPrefixes}
\alias{ref_Honours}
\alias{ref_InterestCategories}
\alias{ref_LordsMembershipTypes}
\alias{ref_LordsRanks}
\alias{ref_OppositionPostDepartments}
\alias{ref_OppositionPosts}
\alias{ref_OppositionRanks}
\alias{ref_OtherParliaments}
\alias{ref_ParliamentTypes}
\alias{ref_ParliamentaryPosts}
\alias{ref_ParliamentaryRanks}
\alias{ref_Parties}
\alias{ref_PartySubTypes}
\alias{ref_PhotoOutputs}
\alias{ref_Statuses}
\alias{ref_titles}
\title{mnis_Reference}
\usage{
ref_AddressTypes()

ref_AnsweringBodies()

ref_Areas()

ref_AreaTypes()

ref_BiographyCategories()

ref_Cabinets()

ref_Committees()

ref_CommitteeTypes()

ref_Constituencies()

ref_ConstituencyAreas()

ref_ConstituencyTypes()

ref_Countries()

ref_Departments()

ref_DisqualificationTypes()

ref_Elections()

ref_ElectionTypes()

ref_EndReasons()

ref_ExperienceTypes()

ref_GovernmentPostDepartments()

ref_GovernmentPosts()

ref_GovernmentRanks()

ref_HonouraryPrefixes()

ref_HonourLists()

ref_Honours()

ref_InterestCategories()

ref_LordsMembershipTypes()

ref_LordsRanks()

ref_OppositionPostDepartments()

ref_OppositionPosts()

ref_OppositionRanks()

ref_OtherParliaments()

ref_ParliamentaryPosts()

ref_ParliamentaryRanks()

ref_ParliamentTypes()

ref_Parties()

ref_PartySubTypes()

ref_PhotoOutputs()

ref_Statuses()

ref_titles()
}
\description{
A series of functions to return reference data. This data is useful for providing parameters for other function calls. The functions do not accept any arguments.
}
\examples{
\dontrun{
x <- ref_AddressTypes()

x <- ref_AnsweringBodies()

}

}
\keyword{mnis}

