% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_tidy.R
\name{mnis_tidy}
\alias{mnis_tidy}
\alias{member_tidy}
\alias{ref_tidy}
\title{Functions to tidy up the variable names returned from the API, and turn dates and datetimes to POSIXct.}
\usage{
mnis_tidy(df, tidy_style = "snake_case")

member_tidy(df)

ref_tidy(df, tidy_style)
}
\arguments{
\item{df}{The tibble to tidy.}

\item{tidy_style}{The style to tidy the tibble with.}
}
\description{
Functions to tidy up the variable names returned from the API, and turn dates and datetimes to POSIXct.

member_tidy

ref_tidy
}
