% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnis_tidy.R
\name{mnis_tidy}
\alias{mnis_tidy}
\alias{ref_tidy}
\alias{constituency_results_tidy}
\alias{tidy_bom}
\title{mnis_tidy}
\usage{
mnis_tidy(x)

ref_tidy(x)

constituency_results_tidy(results, details)

tidy_bom(x)
}
\arguments{
\item{x}{The dataframe to tidy}

\item{results}{The data frame to tidy}

\item{details}{The list to tidy}

\item{x}{The GET returned from call to API}
}
\description{
Internal tidying functions
}
